Remove unneeded file in dm thread list
