+++ mobile/components/ThreadFeed/index.js
@@ -140,6 +140,8 @@ class ThreadFeed extends React.Component<Props, State> {
             loadingIndicator={<Text>Loading...</Text>}
             fetchMore={this.fetchMore}
             hasNextPage={threadConnection.pageInfo.hasNextPage}
+            refetching={this.props.isRefetching}
+            refetch={this.props.data.refetch}
             {...flatListProps}
           />
         </View>
